ExplorationLLM/Research/tagE/pyproject.toml
[tool.poetry]
name = "tage"
version = "0.1.0"
description = ""
authors = []
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pyyaml = "^6.0.1"
torch = "^2.1.1"
numpy = "^1.26.2"
pandas = "^2.1.4"
tqdm = "^4.66.1"
transformers = {extras = ["torch"], version = "^4.36.0"}
sentencepiece = "^0.1.99"
datasets = "^2.15.0"
evaluate = "^0.4.1"
scikit-learn = "^1.3.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
